pyproject.toml
[project]
name = "ap2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.57.1",
    "anyio>=4.9.0",
    "langfuse>=3.1.3",
    "openai>=1.95.0",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.1",
    "rich>=14.0.0",
    "tenacity>=9.1.2",
    "trio>=0.30.0",
    "typer>=0.16.0",
    "types-pyyaml>=6.0.12.20250516",
]

[tool.ruff]
target-version = "py310"  # Updated to Python 3.10+ to support match statements
line-length = 88

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "D",   # pydocstyle
    "N",   # pep8-naming
]
ignore = [
    "D100",    # Missing docstring in public module
    "D101",    # Missing docstring in public class
    "D102",    # Missing docstring in public method
    "D104",    # Missing docstring in public package
    "D107",    # Missing docstring in `__init__`
    "D105",    # Missing docstring in magic method
    "D203",    # 1 blank line required before class docstring
    "D212",    # Multi-line docstring summary should start at the first line
]
exclude = [
    ".git",
    ".venv",
    "venv",
    "__pycache__",
    "build",
    "dist",
]

[tool.ruff.lint.isort]
known-first-party = ["speech_prep"]
force-single-line = false
force-sort-within-sections = true

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.mypy]
python_version = "3.10"  # Updated to match ruff target version
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
strict_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
warn_unreachable = true
disallow_any_generics = false
disallow_subclassing_any = false
explicit_package_bases = true
namespace_packages = true
exclude = [
    "venv/",
    ".venv/",
    "build/",
    "dist/",
]

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
disallow_incomplete_defs = false
disallow_untyped_decorators = false

[dependency-groups]
dev = [
    "isort>=6.0.1",
    "mypy>=1.16.1",
    "pytest>=8.4.1",
    "pytest-mock>=3.14.1",
    "ruff>=0.12.3",
]
